Implement local mode for Commodore
This allows easier local development and disables all interactions with
SYNventory and remote Git repositories.  To use local mode run commodore
compile with the --local flag. The flag takes a Kapitan target as its
argument.

Local mode will try to find all the Commodore components in
`dependencies` but it will not update symlinks for component classes or
update component versions.