Implement committing catalog changes
This commit implements committing the new version of the rendered
cluster catalog to the catalog output repository. The implementation
first deletes the contents of the git working tree directory and then
copies over the compiled catalog. With this method, resources that are
removed in the cluster configuration are also removed in the catalog
repository.