Use click.style and click.echo for colorizing output
Click provides click.echo() which detects whether the output is rendered
on a terminal or into a file, and smartly decides whether to apply
styling which is easily added with click.style(), or the convenience
wrapper click.secho().

This commit colorizes the catalog diff output and bolds the "title"
messages for each stage of the compilation process.