Fix version checkout to work with commit SHAs and tags
The fix in commit 166cbb9 broke the version checkout code for commit
SHAs and tags, as the fix always prepended the provided Git ref with
"remotes/origin" to allow checking out branches that only exist on the
origin remote.

This commit first tries to find a commit for just `ref`, and if that
fails, tries to find a commit for `remotes/origin/{ref}`. This restores
the option to provide tags or commit SHAs as component versions.