Implement -v as counted option and use in clean()
This commit implements -v/--verbose as a counted option to control the
verbosity of Commodore's output.  This commit defines "debug mode" as at
least one instance of -v or --verbose.

This commit uses the verbosity control in the clean() method to print
more information when the option is given enough times to enable debug
mode.

It is not yet clear whether we will continue with this approach of
a counted verbosity option, or if we will switch to a more K8s-like
verbosity option which takes an argument, and which defines different
verbosity levels (1,2,4,8,...) that can be passed (e.g. debug would be
--verbose=8).