Change Commodore to use new, minimal /targets API response
The API now returns a JSON response which only contains the
configurations that are registered in the inventory service when the
cluster registers itself.  Those are cluster_distribution, cloud_type
and cloud_region.

Because we do not have any API endpoints that return YAML after this
change, this commit also reverts commit 6f34ef2 which allowed for
handling YAML responses from the API.