Implement built-in postprocessing filters
After introducing postprocessing to deal with rendered Helm charts where
namespaced Kubernetes objects didn't contain the namespace field, we
quickly saw that most of the components which render Helm charts have to
patch the namespace into the rendered Kubernetes objects.  A first step
was to move a lot of the patching logic into commodore.libjsonnet.

However, this postprocessing step is common enough that we introduce the
concept of built-in filters, which can be used to postprocess Kapitan
output without having to write any Jsonnet in the component.

This commit refactors the postprocessing submodule in Commodore further
to allow different filter types, and provides a first built-in filter
`helm_namespace`.

The new filter definition has a new mandatory field `type`. This field
indicates which type of filter the component wants to call.  Valid
values for the field are `jsonnet` and `builtin`.

"Old-style" filters using custom Jsonnet bundled with the component are
still fully supported as `type: jsonnet`, but Commodore emits a warning
for components which do not specify the filter type in
`postprocess/filters.yml`.

For `builtin` filters Commodore uses the field `filter` as the name of
the builtin filter, and supports an extra optional field `filterargs`.
For `filterargs` Commodore expects a dictionary.  Values in the
`filterargs` dict can use Kapitan syntax to refer to inventory values
(with `${path:in:parameters}`.

Builtin filters are themselves implemented as Jsonnet, and the
implementations can be found under `filters/` in the Commodore
repository.