Switch from ruamel.yaml to PyYAML
Kubernetes uses YAML 1.1 but doesn't recognize the `%YAML 1.1`
directive.
`ruamel.yaml` defaults to YAML 1.2 and doesn't have a simple way to
configure the dump functionality to produce YAML 1.1 without adding
`%YAML 1.1` in the output.

We switch to PyYAML, as that implementation is YAML 1.1 compliant which
avoids compatibility issues between Kubernetes and Commodore.