Move compiled manifests into manifests subdirectory in catalog
This change is done as preparation for enabling secret revealing in
Flux. To use `kapitan refs --reveal` in Flux, the Kapitan ref
definitions need to be available at reveal time. These will be stored in
the `refs` subdirectory in the catalog. Additionally, Flux will need
extra configuration in `.flux.yml` in the root of the catalog
repository.