Reshuffle include hierarchy temporarily to allow incremental testing
To allow incremental testing of components for OpenShift 4 without
having to wait for the pending separation of component defaults and
hierarchy values, the include hierarchy has been temporarily changed to
* global.common
* global.<kubernetes-distribution>
* global.<cloud-provider>
* <customer>.<cluster>

This may get reverted again once value and defaults separation is
implemented. This is in preparation for moving the bulk of component
includes to `global.kubernetes` which is included by `global.rancher`
but not by `global.openshift4`.

Relates: SYN-22