Allow reading API token from file
This commit changes Commodore to optionally read the API token from a
file. If the API token argument (command line or environment variable)
exists as a file, Commodore will read the file contents and use them as
the API token. Otherwise the value of the argument is used as the API
token. This allows using the service account's token when running
Commodore on Kubernetes.

Relates: SYN-21