Fix check whether API token is a file
While illegal characters are not a problem for checking whether the
provided API token value is a path, pathlib fails to properly handle
long tokens (e.g. JWT), and fails with the error `OSError: [Errno 36]
File name too long: '...'` when encountering such a file name.

This commit wraps the whole check whether the provided API token value
is a file name in a try-block and silently assumes that any OSError
thrown by pathlib indicates that the provided value is not a file name.