Ensure Commodore binary version is set when pushing image
The imported Github action for pushing the docker image to DockerHub
does not run the `make docker` command and therefore the Commodore
binary version in the pushed image was not configured correctly.

This commit generates the binary version as a Github action step instead
of relying on the Makefile when building the image to be pushed.