Move bundled Jsonnet components to Commodore python module
This allows building a self-contained Commodore python module which can
be installed without having to copy/retrieve any extra files.

This is achieved by moving the Commodore Jsonnet lib and builtin filters
to the commodore subdirectory in the repository, and adding the files in
`commodore/lib` and `commodore/filters` as additional dependencies in
the Poetry pyproject.toml.