Add rich version information for 'latest' Docker image
The `commodore:latest` Docker image which is built from the master
branch should contain enough version information to allow users to
identify exactly which version of the code is executed.

This can be achieved by using the output of `git describe` with
appropriate arguments to generate a version of the form
`<tag>-<commit count>-g<commit sha>`, where `commit count` is the number
of commits since `tag`, and `commit sha` is the shortened SHA sum of the
HEAD at the time of building the image.

We cannot use this version as the Commodore Python module version, since
it doesn't match the restrictions imposed by PEP-440 [1].

Therefore we implement a secondary "Git version" which is patched into
the Commodore code during the Docker build process using the output of
`git describe` collected during the build process.

`commodore --version` will simply output the Python module version if
the module version and the "Git version" match exactly. It will append
the "Git version" in the output if it differs from the Python module
version.

Note that we considered using PEP-440 developmental releases [2] for the
master branch builds, but there's two hurdles for that approach: 1) due
to how versions are ordered, anything newer than the latest tagged
version should be a development release of the "next" version, requiring
us to define what the "next" version is programmatically, and 2) we
cannot embed all the information from `git describe` in a development
version which is simply `X.Y.devN`, where `N` is a non-negative integer.

[1] https://www.python.org/dev/peps/pep-0440/
[2] https://www.python.org/dev/peps/pep-0440/#developmental-releases