Allow postprocessing filters to be disabled based on reclass references
This can be useful if we need to apply a postprocessing filter on a
subchart of a Helm chart, e.g. Postgres for Keycloak.

This commit refactors the existing reclass reference parsing for builtin
filters to be more general-purpose, and applies the reference parsing
and resolution to the whole filter dict.