Refactor postprocessing filter code
Refactor postprocessing filter code in preparation for moving filter
definitions into the inventory. Defining filters in the inventory will
allow us to rely on Kapitan (and reclass) to resolve inventory
references in the filter definitions.

This commit changes some of the postprocessing filter execution
internals to prepare the code base for an implementation which allows
both styles of filter definitions (in the inventory and in
postprocess/filters.yml) in a transitional period until all existing
components have been updated.

This commit does not yet introduce functionality to extract filters from
the inventory, but refactors the internal filter dict contents to reduce
the amount of required fields for filter definitions, especially for
builtin filters.