Add code to run postprocessing filters from inventory
Postprocessing filters can also be defined under
`parameters.commodore.postprocess.filters` in component classes using
mostly the same format as `postprocess/filters.py`.

Notable changes to the filter definition format: Jsonnet filters need to
provide the full path to the directory holding the filter definition
(including the component directory) as `filterpath`. Filters which
require the component name need to set it as `component` in the filter
definition. Filters without `type` are not accepted if defined in the
inventory.