Switch os.system() to subprocess.call()
os.system() returns a number that is a (binary) mix between the syscall
that terminated the program and its exit code. In case of successful
program exit, there is no difference, but in case of error exit, the
result needs to be binary shifted (or masked) to get the actual exit
code.

This branch switches to using the recommended way to call a shell
command in python3 instead (subprocess.call()).

See https://docs.python.org/3/library/os.html#os.system for more
context.