Implement slug validation for `component new`
In https://github.com/projectsyn/commodore/pull/186 we decided to
disallow component slugs with a leading digit. While implementing this
policy, I've suggested to restrict slugs further.

After some further discussion, we decided to restrict component slugs to
the regex `^[a-z][a-z0-9-]+[a-z0-9]$`, when creating components with
`component new`.

This commit implements slug validation in the component templater using
the regex mentioned above, and introduces a number of test cases which
validate the new behavior.