Provide inventory_path in Kapitan's argument cache
Kapitan 0.29.3 contains  [PR#625] where the default inventory path
`./inventory` is removed and the default path is extracted from the
command line argument if necessary.

Since we do not call Kapitan via the CLI we need to manually populate
the argument cache.

This commit introduces a namedtuple `ArgumentCache` which can be
provided as a minimal argument cache to Kapitan's cached module. This
change ensures that our calls to Kapitan's compile_targets method
continue working with Kapitan 0.29.3

[PR#625]: https://github.com/deepmind/kapitan/pull/625