Implement support for component instantiation
This commit implements component instantiation as described in
[SDD0025]. Instances can be defined by users by using the aliasing
syntax using the keyword "as" in entries in the `applications` list.

Commodore does some basic sanity checking and rejects configurations
with conflicting aliases.

The current implementation uses the presence of key `_instance` in the
component's parameters dict as the indicator that the component supports
instantiation, and sets this parameter to the instance alias for aliased
instances. The implementation raises an error if aliases exist for
components which don't have this key defined in their defaults class.

[SDD0025]: https://syn.tools/syn/SDDs/0025-commodore-component-instantiation.html