Switch to `parameters.<component_name>.multi_instance` for declaring instance support
Components now need to set `parameters.<component_name>.multi_instance`
to true to signal that they support instantiation.

The meta-parameter `_instance` is now available as
`parameters._instance`, which shortens its use in reclass references.

The parameter `_instance` is always set to the instance name, which is
equal to the component name for non-aliased components.