Always create targets from component aliases
This commit addresses the issue reported in #268, where we found that if
a component is only included in aliased forms, the plain component is
rendered to the catalog in addition to the aliased versions.

This commit changes component discovery to handle unaliased components
as if they were aliased to the component name itself.

This change in discovery allows us to only generate targets for aliases
which correctly doesn't create a target for the unaliased version of a
component if it's only included in aliased form.

Fixes #268.