Merge component_versions into components in `params.cluster`
Previously, for #284, we've merged component_versions into components in
each Kapitan target.

This led to counter-intuitive behavior when trying to override versions
in `components`, if the same component also has a version override in
`component_versions`, as the override in `component_versions` would
always have precedence over overrides in `components` regardless of
their relative order in the hierarchy.

This component moves the merging of `component_versions` into
`components` to the class `params.cluster` which has the lowest
precedence of all classes.

This ensures that configuration in `components` has precedence over
configurations in `component_versions` in all cases, while still
allowing configurations which use `component_versions` to function.