Correctly handle component versions which are tags
Tags show up as heads in the remote, but without the remote name as
prefix. Previously the code would then try to checkout a branch with the
name of the tag.

This commit makes the assumption that any heads in a remote which don't
have the remote name as a prefix are tags, and handles checkout
accordingly.

To keep the things simple, tags are checked out in detached HEAD state.