Add command line options to override global and tenant repo revisions
These options are intended to be used to test changes to the
global/tenant repo without relying on local compilation.

When either of the repo revisions are overridden using the command line
flag, the resulting catalog cannot be pushed.

If a cluster catalog should actually be compiled and pushed using a
different revision of the global or tenant repo, set the override in
Lieutenant.