Simplify `component new` and `component delete`
Remove the code which adds and removes the new component in the
inventory and `jsonnetfile.json`.

Local mode now fully supports component discovery, so the required
symlinks and entries in `jsonnetfile.json` are created automatically by
`catalog compile --local` after the user has added the new component(s)
to the `applications` array.

We keep the call to `delete_component_symlinks` in `component delete` to
avoid having dangling symlinks as the contents of
`inventory/classes/defaults` and `inventory/classes/components` are not
reset in local mode.