Reset and recreate component symlinks in local mode
With the new implementation of local mode we can recreate the component
class symlinks in the inventory on each run instead of relying on the
pre-existing symlinks.

This also allows us to remove the call to `delete_component_symlinks` in
`component delete`, as the next run of `compile` will always delete and
recreate all class symlinks.