Implement support for deprecating components
This commit introduces a new component parameter `_metadata` which is
defined as constant in the component template. When adding the parameter
to existing components users are responsible to ensure that they mark
the parameter as constant (with the `=` prefix).

Further, this commit changes Commodore will print out deprecation
notices for all components for which
`parameters.<component-name>._metadata.deprecated` is `True`.