Fix releases to show "x commits to master since this release"
This notice is only shown when the release is created from a branch, and
not if it's created from a commit id or tag, see [this discussion].

This commit ensures that all releases are created from branch "master"
and therefore have the message showing how many commits have been added
to "master" since the release.

Additionally, switch to https://github.com/ncipollo/release-action as
the release action, since `actions/create-release` is deprecated.

[this discussion]: https://github.community/t/x-commits-to-branch-since-this-release-is-missing-part-2/1806