Implement optional `output_dir` postprocess filter argument
Previously Commodore postprocessing filters had the hard-coded
assumption that each component's output would appear in
`compiled/<component-name>`. This is not necessarily true, e.g. when an
instance-aware component creates individual ArgoCD apps and folders for
each instance.

This commit introduces an optional paramter `output_dir` for
postprocessing filter definitions which allows users to override the
default directory in `compiled` in which Commodore applies
postprocessing filters for the component.