Implement option to specify target alias for `commodore component compile`
Previously we faked "instances" when compiling components by just
setting `parameters._instance`, but kept the target for component
compilation as the component name. This leads to differences in the
resulting files as seen by the postprocessing step depending on whether
a component is compiled independently or as part of a cluster catalog.

This commit introduces an additional command line option to specify the
target component alias to use when compiling a single component with
`commodore component compile`.

When this option is given, `component compile` correctly configures an
aliased component and creates a target for the alias instead of the
component. This ensures that postprocessing sees identical paths in
`compiled` regardless of whether the component is being compiled as part
of a cluster catalog or independently.