Import stripped-down inventory rendering code
This is roughly the code implemented for
https://github.com/projectsyn/commodore-defaults-commentator to render a
partial inventory outside of catalog compilation.

This version is simplified and doesn't currently provide support for
actually extracting annotations, since it relies on Kapitan's bundled
reclass which doesn't have support for keeping location information in
the rendered reclass inventory.

The implementation tries to provide required parameters and facts which
are used by VSHN's global defaults but aren't set to a default value in
the defaults repo itself.