Refactor `commodore inventory components`
The command now takes an arbitrary number of values files which are
used as additional reclass classes instead of individual parameters for
specifying the distribution, cloud and region for which to render the
global defaults repo.

This allows users to both provide facts (e.g. distribution, cloud,
region) and additional required parameters for a given global defaults
repository.

The command also takes an optional parameter which controls whether
reclass raises an error on missing classes. This can be useful in cases
where the input to the command should result in a reclass hierarchy
where no classes are missing.

To make this work, the command also creates an empty class for cluster
`c-bar` in tenant `t-foo`. If users overwrite the cluster and tenant ids
via the values mechanism, the `--no-allow-missing-classes` parameter may
raise errors if the provided cluster class doesn't exist.