Properly handle 403 replies from Lieutenant
We previously assumed that any 4xx replies from Lieutenant would be
Lieutenant errors, and would therefore contain key `reason` in the
response.

However, if you try to compile a cluster to which you don't have access
with a restricted Lieutenant API token (e.g. a cluster or tenant SA
token), you get a 403 response which doesn't contain a `reason` field.

This commit correctly handles this case, and produces a legible error
message in this case.

Fixes #367