Use `GitRepo.clone()` when fetching cluster catalog
This ensures that Commodore doesn't fail when cloning an empty cluster
catalog repo for the first time.

Additionally, this commit adds tests for `GitRepo.clone()`, and
`catalog.fetch_catalog()` for the "unusual" cases (cloning an empty
repo, and cloning a non-existent repo).

The common case (cloning a cluster catalog repo which already has
content) is covered by the integration test.