Add make targets `gen-golden-all` and `golden-diff-all` in component template
These targets are written to use a static list of test instances in make
variable `test_instances`. Since the initial template only has a single
test instance, we just hard code the default value of the variable in
the component template.

We use a generic recursive make target based on the `test_instances`
make variable to implement the new targets. This allows us to reuse the
existing `golden-diff` and `gen-golden` targets.

The new targets are only added for components which have matrix tests
enabled (command line flag `--matrix-tests`).