Raise error when compiling a cluster which uses removed reclass variables
We raise an error if we find any occurrences of removed reclass
variables (cf. #460) in the cluster's global defaults, tenant repo, or
any enabled components.

The global and tenant repo are linted before we try to fetch or register
components, so that we always are able to produce nice errors instead of
a reclass stack trace if the inventory contains usages of removed
parameters.

For the components, we check the `defaults.yml` and
`<component-name>.yml` of each included component after fetching or
registering the cluster's components.