Move frequently used pytest fixtures to `tests/conftest.py`
This allows us to use them in all tests without redefining them in each
test file. Additionally, the commit normalizes the naming (having a
fixture `data` is not particularly helpful when trying to figure out
what it does).

Additionally, we adjust the tox config to run tox in an empty temp dir.
This ensures we don't get import conflicts, since pytest also changes
the python import paths when it detects a `conftest.py`.