Change target class to be provided as a path
Some packages may require that some configuration is provided in an
additional class file, for example cluster facts may be used to include
additional classes by the package.

Generally, users should define test cases in `tests/` in the package
directory and have each test case class include the desired target class
of the package in the test class's `classes` array. Note that relative
includes are always interpreted relative to the class containing the
include. Reclass supports the prefix `..` to include a class from the
parent directory.