Change `Component` and `Package` implementations to use `MultiDependency`
This commit changes how Commodore handles dependencies to implicitly
assume that any set of dependencies may share a repository. This changes
the disk layout for the checked out dependencies to use `git worktree`
with a bare "main" clone of the repository stored in
`dependencies/.repos/git.host/path/to/repo.git` where the path under
`dependencies/.repos` is constructed from the remote URL.

Addtionally, this commit updates all tests which interact with component
or package repos to work with the new structure.