Create new components and packages as Git worktree checkouts
Only create worktrees when creating components or packages in a
Commodore working tree (i.e. as subdirectories of `dependencies/`)

We also change the generic dependency template rendering to always
render the template in a temp directory and moving the result into the
final location, because we can't make an existing directory a Git
worktree and cookiecutter doesn't support rendering a template into an
existing directory, so we need to create the worktree first, then copy
the templated dependency into it.