Deduplicate requested test cases for `package new` and `package update`
We previously didn't deduplicate the set of test cases requested when
creating or updating a package. This can lead to noisy updates when
users specify a new test case which already exists in the package.

This commit adds logic to remove duplicates in the getter for the list
of test cases.  We deduplicate the list in the getter rather than the
setter, to avoid introducing invisible duplicates if users use
`append()` on the returned list.