Fix `package update` commit message to use correct template commit SHA
We previously read the template commit SHA from the package's
`.cruft.json` before actually performing the update. However, this
commit SHA wouldn't match the actually commit SHA of the actual
template commit when updating a package to a new template version.

This commit changes the field `PackageTemplater.template_commit` to a
property and reads the value from `.cruft.json` at the time the field is
read. This ensures that we always read the correct commit SHA when we
commit template changes in `package update`.