Add flag for providing fallback dynamic facts for `catalog compile`
We support providing dynamic facts as `--dynamic-fact key=value` in
command `catalog compile`. The key can specify a nested key as
`path.to.key`, and the value can optionally be parsed as JSON if it's
prefixed with a literal `json:`. Facts specified later on the command
line take precedence over previously specified facts. Nested facts will
not be considered if one of their key segments (except the leaf) already
holds a non-dictionary value.

Dynamic facts provided on the command line are only used for catalog
compilation when the Lieutenant API response for the cluster which is
getting compiled doesn't contain field `dynamicFacts`.

This enables us to cleanly support compilation of a cluster catalog
before Steward is deployed on the cluster.