Update integration test to use mocked HTTP responses
Previously we mocked the function `load_cluster_from_api()`, but now
that we're using `responses` for mocking HTTP responses anyway, we can
just mock the expected API calls (for the cluster and tenant) and allow
any other requests to pass through.

This should make the integration test slightly closer to a "real"
`catalog compile`, as the normal code for parsing the Lieutenant
response is used.