Add logic to try refreshing tokens in `login()`
We implement OIDC token refresh if our access token is expired, but we
still have a valid refresh token in the tokencache for the current
Lieutenant URL. If the token refresh succeeds, we short-circuit
`login()` without performing a browser-based login.

If the token refresh fails for any reason, we fall back to browser-based
login.

Closes #494