Fix test for `login.login()`
We accidentally made the test a no-op with the early return in the
function if `config.api_token` is set.

We need to explicitly clear the api_token field in the `config` test
fixture to actually test the login flow now.

To catch such errors in the future we now also assert that the mocks for
`webbrowser.open` and `commodore.tokencache.save` are called exactly
once.