Add new command `package sync`
This command reads a list of GitHub repositories containing Commodore
config packages from a file provided on the command line and runs
`package update` for each repository. It creates or updates PRs
containing the latest template changes for each listed package.

We ensure that we don't create PRs too rapidly by sleeping for 1-2
seconds after creating or updating a PR.