Fix console output for `package sync`
Previously we returned `None` from `ensure_pr()` when doing a dry-run
sync or on an error. However, we printed the regular success message on
None returns. This creates a output message along the lines of "PR for
package organization/package successfully None".

This commit refactors how we generate and print the final message for
each package. Instead of returning "created", "updated" or None from
`ensure_pr()`, we return the preformatted message which can just be
printed in `sync_packages()`.