Update suggested Commodore in Docker helpers
We need to mount the host working directory into the same location in
the container so that `git worktree` based checkouts of components and
packages are valid both on the host and in the container.

This needs a identity-mapping of the working directory because Git uses
the absolute path to the git directory when creating a worktree
checkout.