Deprecate parameter `--repo-directory` for `component compile`
We don't need users to provide the location of the repo directory in
`component compile` since we strictly speaking don't need the component
to even be in a Git repository for standalone compilation. We  use
GitPython's `search_parent_directories=True` to try to find the repo in
which the provided target dir is located, but fall back on just using
the indicated target directory as the component directory for standalone
compilation.

If we find a Git repository holding the component, we compute the
component subpath based on the relative location of the target directory
to the repo working tree directory. In this case, we use the working
tree directory of the repo as the component "target directory" so that
standalone compilation of a component stored in a repo subdirectory is
as close as possible to compiling the component in a cluster catalog.

We don't remove the `--repo-directory` parameter, but print a
deprecation warning when users provide it, and ignore the provided value
in the implementation.