Ensure that there's no chance that test cases compete for ports
We added more tests for `login.py` a while ago. Those tests introduced a
race condition between test cases as we have two different tests which
try to bind to port 18000 to test the OIDC callback handler.

This commit makes the port configurable, and uses a different port for
the callback handler unit tests.