Update Commodore-in-Docker for Linux to enable GPG-signing commits
We update the Commodore-in-Docker command for Linux to ensure that Git
commits can be GPG-signed in the container when the user's `.gitconfig`
contains `commit.gpgsign=true`. To allow this, we mount the user's
`pubring.kbx` and the user's GPG `agent-extra-socket` in the container,
if GPG is available and configured.

We check for availability by looking for the `gpgconf` command on the
host system, and for configuration by looking for the user's
`pubring.kbx`.

Please note that the Commodore-in-Docker command for macOS doesn't
provide out-of-the box support for GPG-signing commits in the container
at this time.