Fix dependency spec linting for dependencies in subpaths
In Commodore v1.4, we introduced support for storing dependencies in
subpaths of repositories. However, we didn't update the dependency
specification linting to allow the new key `path` which is used to
indicate the subpath.

This commit updates the dependency specification linting to allow key
`path` in component and package specifications.