Fix GitRepo initialization for dependencies without backing MultiDependency
We don't always create a `MultiDependency` for `component compile`, so
we need to ensure `Component.repo` works correctly even if
`Component._dependency` is `None`.

This commit ensures that we fall back to using `None` for the Git author
name and email if we don't have a `MultiDependency` for a component.