Add option to ignore template commit id changes in `Templater`
Add a optional argument for `Templater.update()` which configures a
custom diff function which suppresses `.cruft.json` diffs which only
contain a changed template commit id.

This allows us to change dependency sync to not create PRs for updates
which only modify the template commit id for template updates which make
changes in a part of the template which only applies to a few
dependencies.