Implement support for taking over library prefix of deprecated components
Commodore will verify that either the component who's prefix we're
trying to take over isn't present in the cluster configuration (in this
case the alias is just a regular alias) or that the component is
deprecated, and has nominated us as their replacement component. Library
prefixes of non-deprecated components, or deprecated components which
don't specify a replacement can't be taken over.