Fix crash in `inventory lint` when top-level `parameters` is not an object
Until now, we didn't check whether a top-level key `parameters` is
actually a dict, which caused a crash in the linter when a non-inventory
file (e.g.  an OpenShift template) has such a top-level key.

This commit simply skips files which have a top-level key `parameters`
which isn't a dict.

Part of #659