Implement ignore patterns for `inventory lint`
Patterns can be provided either on the command line in
`--ignore-patterns` or in a `.commodoreignore` file in the linting
target path. Patterns given on the command line apply for each target
path, while `.commodoreignore` applies only for the target path in which
the file is present.

Patterns are expanded recursively using Python's `glob` library.  You
can use `*`, `?`, and character ranges expressed as `[]` with the usual
semantics of shell globbing, and `**` to indicate an arbitrary amount of
subdirectories. Patterns which start with `/` are treated as anchored in
the target path, while other patterns are treated as matching any
subpath in the target path.

Part of #659