Update dependency fetching args to match new Kapitan version
Kapitan has changed the dependency fetching arguments to
`compile_targets`.

This commit updates our compile helper to set the new arguments so that
we only fetch dependencies when Commodore's `--no-fetch-dependencies`
isn't given.

Without this change, Kapitan's dependency fetching was active regardless
of whether `--no-fetch-dependencies` was given as an argument to
`commodore catalog compile`.