Only print a message when the OIDC login URL can't be opened automatically
Until now, we unconditionally printed a message containing the OIDC
login URL whenever Commodore needed to fetch a new OIDC token. The
`webbrowser.open()` function returns a boolean which indicates whether
the URL was opened successfully.

This commit checks the return value from `webbrowser.open()` and only
prints a message if the URL couldn't be opened. This reduces the chances
of this message messing up the terminal output when autocompletion is
used (cf. #722), and reduces noise in general.