Update `component compile` instance test to access instance parameter
Until now, we didn't cover the case that a component test case for a
multi-instance-aware component wants to overwrite a component parameter
in the instance's parameter key.

This case also didn't work correctly, since we never added the logic to
merge the instance parameters into the component parameters in the
target generated by `commodore component compile`.