Factor out target content generation from `render_target()`
This commit pulls the target object generation out from
`render_target()`. The new `generate_target()` function expects to be
passed a list of classes to include in the target, but contains all the
logic which creates the target `parameters` contents which was
previously in `render_target()`.

In contrast to `render_target()` which does some work that's unnecessary
for `component compile` and doesn't allow injecting the additional test
classes, the `generate_target()` function just encapsulates the
parameters logic which is identical for `catalog compile` and `component
compile`, and simply puts the provided `classes` array in the resulting
object.