Use `generate_target()` in `component compile`
We replace the manually crafted target object in
`_prepare_kapitan_inventory()` with a call to `generate_target()`,
providing the same manually crafted classes list as before.

This fixes the bug in `compnent compile` that parameters overwritten in
the instance's key don't get picked up by `component compile`, because
we now get the same `parameters` logic as for `catalog compile`.