Provide cluster and tenant display names in `component compile` and `package compile`
We manually generate a `cluster/params.yml` in `component compile` and
`package compile`. This logic hasn't been updated to provide values for
`${cluster:display_name}` and `${cluster:tenant_display_name}` in #311.

This commit updates `component compile` and `package compile` to provide
dummy values for those variables and adjusts the component compilation
tests to verify that these variables are provided.